hits/requirements.txt
Django>=3.0
djangorestframework>=3.12
psycopg2-binary>=2.8